pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    { name = "Vasilii Stakrotckii", email = "142693897+inferno681@users.noreply.github.com" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.7"
sqlalchemy = "^2.0.37"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
httpx = "^0.28.1"
uvicorn = "^0.34.0"
apscheduler = "^3.11.0"
pytz = "^2024.2"
pydantic-settings = "^2.7.1"
pyyaml = "^6.0.2"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
types-pytz = "^2024.2.0.20241221"
types-pyyaml = "^6.0.12.20241230"
flake8 = "^7.1.1"
isort = "^5.13.2"
wemake-python-styleguide = "^1.0.0"

[[tool.mypy.overrides]]
module = ["apscheduler.*"]
ignore_missing_imports = true
